feat: Hide app_company_id field on index and when creating/updating user
This commit adds functionality to hide the app_company_id field from the index view and when creating or updating a user. The field is still visible on detail view and only accessible by users with the 'super_admin' role. Additionally, after creating or updating a user, if the user has a company_id, they are assigned the 'company_admin' role and their app_company_id is set to match their company_id. If the user does not have a company_id, they are removed from the 'company_admin' role and their app_company_id is set to null.