feat: Add URL field to Zone resource
This commit adds a new URL field to the Zone resource in order to display and link URLs. The URL is now wrapped if it exceeds 75 characters, and new lines are replaced with `<br>` tags for proper formatting. The URL is displayed as a clickable link with dark blue color, and must start with either "http://" or "https://".