feat: Add functionality to mark ticket as read after sending email reply
This commit adds a new feature to the TicketAnswerViaMail.php file. After sending an email reply, the ticket is now marked as read by setting the "is_read" property to true. This change improves the user experience by providing visual feedback on whether a ticket has been read or not.