feat: feature Vip
feat: add ticket ID to notification data

Add the ticket ID to the data payload of the notification sent in the ProcessTicket job. This allows for better tracking and identification of tickets in the notification system.

feat: add sound to push notification

Add the 'sound' parameter to the push notification payload in the ProcessTicket job. This allows for a custom sound to be played when receiving the notification.

feat: Add push notifications for VIP tickets

This commit adds functionality to send push notifications for VIP tickets. The `ProcessTicket` job now accepts an additional parameter `$event` to determine the type of event (created or updated). If the event is 'created', a push notification is sent to dustyman users. If the event is 'updated' and the ticket status is 'done', a push notification is sent to the VIP user.

feat: Add logging for ticket creation and status update

- Added logging for ticket creation in the ProcessTicket job
- Added logging for ticket status update in the ProcessTicket job

chore: Refactor TicketCreated mailer to include image attachment

The TicketCreated mailer in the app/Mail/TicketCreated.php file has been refactored to include an image attachment if the ticket object contains an image. This enhancement allows for sending images along with the ticket creation email.

chore(mail): add VIP prefix to ticket email subject

This commit adds a VIP prefix to the ticket email subject if the user has the 'VIP' role. This helps differentiate VIP tickets from regular tickets in the email subject line.

Changes made:
- Added logic to check if the user has the 'VIP' role
- Added a variable to store the VIP suffix ('VIP ')
- Modified the email subject to include the VIP suffix

chore: Update TicketCreated email template

- Fix role check for VIP users
- Update email subject to include VIP prefix

fix: remove unnecessary space in VIP suffix

The commit fixes an issue where an unnecessary space was added to the VIP suffix in the email subject.

feat: Add filter for ticket status in index method

Added a condition to filter out tickets with the 'done' status in the index method of the TicketController. This allows users to view only active tickets.