feat: Add SendPushNotification action to Nova
This commit adds a new file `SendPushNotification.php` to the `app/Nova/Actions` directory. The file contains the implementation of the `SendPushNotification` action class, which is responsible for sending push notifications to users.

The `SendPushNotification` action class has a `handle` method that takes in an instance of `ActionFields` and a collection of users as parameters. It retrieves the FCM tokens from the users and sends a push notification using Larafirebase package. The result of the notification sending is checked, and an appropriate message is returned based on the response status.

Additionally, this commit modifies the `User.php` file in the `app/Nova` directory. It adds an instance of the `SendPushNotification` action to the list of actions available for users in Nova. The action is only visible to super admins.

These changes enhance the functionality of Nova by allowing super admins to send push notifications to users directly from their user profiles.