chore: Add command to update collected tickets status oc: 3696 (#17)
* chore: Add command to update collected tickets status

This commit adds a new console command `UpdateTicketCollectedStatus` that updates the status of tickets that have been collected for more than one day to "done". The command is scheduled to run daily at 22:00.

feat: Add 'collected' status to TicketStatus enum and Nova Ticket resource

This commit adds a new status 'collected' to the TicketStatus enum in the app/Enums/TicketStatus.php file. It also includes the corresponding translation for 'collected' in the lang/vendor/nova/it.json file. Additionally, the Nova Ticket resource in app/Nova/Ticket.php has been updated to handle the new 'collected' status by assigning it a specific color.

feat: Update push notification handling

- Refactored code in ProcessPushNotification.php to use Larafirebase's new methods for setting title, body, sound, and additional data.
- Updated code in ProcessTicket.php to use Larafirebase's new methods for setting title, body, sound, and additional data.
- Modified code in SendPushNotification.php to use Larafirebase's new methods for setting title, body, sound, and additional data.

* chore: Remove default sound from push notifications

The code changes remove the line of code that sets the default sound for push notifications in three different files: ProcessPushNotification.php, ProcessTicket.php, and SendPushNotification.php.

chore: Update ticket collected status logic

- Refactored the code to calculate the correct number of days ago, excluding weekends.
- Updated the query to fetch tickets with a 'collected' status that were last updated more than 3 business days ago.

* chore: Update ticket status command and related code

- Renamed the signature of the command from 'tickets:update-collected' to 'tickets:update-execute'
- Updated the description of the command to reflect the change in ticket status
- Modified the query in the handle method to filter tickets with status 'execute' instead of 'collected'
- Updated the schedule in Kernel.php to use the new command name
- Removed 'collected' status from TicketStatus enum and related code references
- Modified ProcessTicket job to check for ticket status as 'execute' instead of 'collected'
- Removed 'collected' option from headerFields method in Ticket model
- Updated Italian translation file (it.json) by removing translation for 'collected'

These changes ensure that tickets with a status of 'execute' are properly handled and updated.

* chore: Update SendBatchNotification to include additional data in push notification

This commit updates the SendBatchNotification class to include additional data in the push notification. The additional data includes a 'page_on_click' field with the value '/push-notification'. This change allows for more customization and functionality in the push notifications sent by the application.