feat: Add ExportableModel class for exporting models to Excel
This commit adds a new class called ExportableModel, which is an abstract class that extends the Eloquent Model. The ExportableModel provides a method called exportToExcel, which can be used to export data from the model to an Excel file.

The exportToExcel method accepts parameters such as filename, columns, headerLabels, and queryCallback. It uses the Maatwebsite\Excel package to generate the Excel file.

The ExportableModel also includes some helper methods for formatting the exported data, such as WithMapping, WithHeadings, WithStyles, ShouldAutoSize.

This commit also updates the Waste model to extend the ExportableModel instead of the Eloquent Model. Additionally, it adds a new Nova action called ExportWasteToExcel for exporting waste data to an Excel file.

Lastly, this commit adds dependencies on the maatwebsite/excel package in composer.json and updates the composer.lock file accordingly.