feat: Add ExportableModel class for exporting models to Excel oc: 4314 (#18)
* feat: Add ExportableModel class for exporting models to Excel

This commit adds a new class called ExportableModel, which is an abstract class that extends the Eloquent Model. The ExportableModel provides a method called exportToExcel, which can be used to export data from the model to an Excel file.

The exportToExcel method accepts parameters such as filename, columns, headerLabels, and queryCallback. It uses the Maatwebsite\Excel package to generate the Excel file.

The ExportableModel also includes some helper methods for formatting the exported data, such as WithMapping, WithHeadings, WithStyles, ShouldAutoSize.

This commit also updates the Waste model to extend the ExportableModel instead of the Eloquent Model. Additionally, it adds a new Nova action called ExportWasteToExcel for exporting waste data to an Excel file.

Lastly, this commit adds dependencies on the maatwebsite/excel package in composer.json and updates the composer.lock file accordingly.

* chore: Update subproject

- Deleted `ExportableModel.php` file
- Created `ModelToExcel.php` file
- Removed `ExportableModel` trait from `Waste` model
- Renamed `ExportWasteToExcel.php` to `ExportToExcel.php`
- Updated the code in `ExportToExcel.php` to use the new `ModelToExcel` class and added parameters for export models, columns, relations, and filename
- Updated the code in `Waste.php` to use the new `ExportToExcel` action with appropriate parameters

* chore: Add ModelToExcel class for exporting Eloquent models to Excel

This commit introduces a new class called ModelToExcel, which allows easy export of Eloquent model data to Excel files. The class supports custom columns, relationships, and styles. It implements several interfaces such as FromCollection, WithHeadings, WithStyles, WithMapping, and ShouldAutoSize.

The ModelToExcel class can be instantiated with a query builder for the model, an array of columns to export (with optional custom headers), an array of relations to include, and a callback function for custom styling. If no columns are specified, all model fields will be exported.

The commit also includes updates to the ExportToExcel action in the Nova directory. The action now accepts the ModelToExcel instance with the necessary parameters for exporting data.

Additionally, the Waste resource in the Nova directory has been updated to use the new ExportToExcel action with appropriate column and relation configurations.

This change improves the functionality of exporting Eloquent models to Excel files in Laravel applications.

* chore: Update subproject

- Added new file `ExportFormat.php` to the `app/Enums` directory.
- Renamed file `ModelToExcel.php` to `ModelExporter.php` in the `app/Models` directory.
- Renamed file `ExportToExcel.php` to `ExportTo.php` in the `app/Nova/Actions` directory.
- Updated references and imports accordingly.
- Added new field for selecting export format in the Nova resource file.
- Updated dependencies in the composer.json file.

* chore: Remove unused code and dependencies

- Removed the `MPDF` export format from the `ExportFormat` enum.
- Removed the `ModelExporter::DEFAULT_STYLE` constant and updated its usage in the constructor.
- Updated the default value of `$styles` parameter in the `ModelExporter` constructor to use `self::DEFAULT_STYLE

* chore: Remove unused export formats and update export functionality

- Removed the HTML and DOMPDF export formats from the ExportFormat enum
- Updated the ExportTo action to ensure that the export folder exists before exporting
- Added a method to clear the contents of the export folder before exporting
- Updated the file path for storing the exported file in the ExportTo action
- Deleted the config/excel.php file, as it is no longer needed

* feat: Add ExportDownloadController and update ExportTo action

- Added a new file `ExportDownloadController.php` in the `app/Http/Controllers` directory.
- Implemented the `download` method in the `ExportDownloadController` to handle file downloads.
- Updated the `ExportTo` action in the `app/Nova/Actions/ExportTo.php` file.
  - Removed unused import statement for `File`.
  - Added import statement for `URL`.
  - Modified the logic to generate a unique filename using current timestamp.
  - Changed the storage location and filename when exporting data using Laravel Excel.
  - Generated a temporary signed URL for downloading the exported file.
- Added a new route `/download-export/{fileName}` in the `routes/web.php` file, which maps to the `download` method of `ExportDownloadController`. The route is protected with signed middleware.

These changes enable users to download exported files from Nova.